updated single-cell processing pipeline

+++ utilities/single-cell/single_cell_pipleline.sh
@@ -12,8 +12,9 @@ conda activate $bisulfite_env
 bam_directory='/home/dwk681/workspace/cluster_cells_from_GSE189158_NOMe_HiC/filesFromCluster/bam'
 software_directory='../../bin/softwarefiles'
 chrom_file="../../bin/softwarefiles/hg19.autosome.chrom.sizes"
-filtered_list='filtered_bam_list.txt'
+fragments_file="$bam_directory/hg19_DpnII.txt"
 output_directory='../../projects/single_cell_files'
+filtered_list="$output_directory/filtered_bam_list.txt"
 
 # Create output directory if it doesn't exist
 mkdir -p $output_directory
@@ -24,36 +25,43 @@ min_high_quality_reads=250000
 # Load samtools module
 module load samtools
 
-# Create or clear the filtered list file
-> $filtered_list
+# Check if the filtered list already exists
+if [ ! -f "$filtered_list" ]; then
+  # Create or clear the filtered list file
+  > $filtered_list
 
-# Count the initial number of BAM files
-initial_bam_count=$(ls $bam_directory/sc*.b37.calmd.bam | wc -l)
-echo "Initial number of BAM files: $initial_bam_count"
+  # Count the initial number of BAM files
+  initial_bam_count=$(ls $bam_directory/sc*.b37.calmd.bam | wc -l)
+  echo "Initial number of BAM files: $initial_bam_count"
 
-# Loop through BAM files and filter based on quality
-for bam_file in $bam_directory/sc*.b37.calmd.bam; do
-  echo "Evaluating $bam_file"
-  
-  # Count high-quality reads
-  high_quality_reads=$(samtools view -c -q 30 -f 1 -F 1804 "$bam_file")
-  
-  # Check if the BAM file meets the quality criteria
-  if (( high_quality_reads >= min_high_quality_reads )); then
-    # Extract the identifier and add it to the filtered list
-    identifier=$(basename "$bam_file" .b37.calmd.bam)
-    echo "$identifier" >> $filtered_list
-  fi
-done
+  # Loop through BAM files and filter based on quality
+  for bam_file in $bam_directory/sc*.b37.calmd.bam; do
+    echo "Evaluating $bam_file"
+    
+    # Count high-quality reads
+    high_quality_reads=$(samtools view -c -q 30 -f 1 -F 1804 "$bam_file")
+    
+    # Check if the BAM file meets the quality criteria
+    if (( high_quality_reads >= min_high_quality_reads )); then
+      # Extract the identifier and add it to the filtered list
+      identifier=$(basename "$bam_file" .b37.calmd.bam)
+      echo "$identifier" >> $filtered_list
+    fi
+  done
 
-# Count the number of BAM files in the filtered list
-filtered_bam_count=$(wc -l < $filtered_list)
-echo "Number of BAM files in the filtered list: $filtered_bam_count"
-echo "Filtered list of BAM files has been created."
+  # Count the number of BAM files in the filtered list
+  filtered_bam_count=$(wc -l < $filtered_list)
+  echo "Number of BAM files in the filtered list: $filtered_bam_count"
+  echo "Filtered list of BAM files has been created."
+else
+  echo "Filtered list already exists. Skipping filtering step."
+fi
 
 # Create symbolic links to BAM files in the output directory based on the filtered list
 while read identifier; do
-  ln -s "$bam_directory/$identifier.b37.calmd.bam" "$output_directory/$identifier.bam"
+  if [ ! -L "$output_directory/$identifier.bam" ]; then
+    ln -s "$bam_directory/$identifier.b37.calmd.bam" "$output_directory/$identifier.bam"
+  fi
 done < $filtered_list
 
 # Ensure that bisulfite_env_python_path is set correctly
@@ -70,16 +78,23 @@ for bam_file in $output_directory/sc*.bam; do
   good_reads_bam="$output_directory/${prefix}.good_reads.bam"
   hic_txt="$output_directory/${prefix}.hic.txt"
 
-  # Filter BAM file and convert to juicer_short format
-  samtools view --threads 5 -bh -q 30 -f 1 -F 1804 "$bam_file" > "$good_reads_bam"
-  if [ $? -ne 0 ]; then
-    echo "Error processing $bam_file with samtools."
+  # Skip processing if output files already exist
+  if [ -f "$good_reads_bam" ] && [ -f "$hic_txt" ]; then
+    echo "Output files for $bam_file already exist. Skipping."
     continue
   fi
 
-  "$bisulfite_env_python_path" "$software_directory/sam2juicer.py" -s "$good_reads_bam" -f "$bam_directory/hg19_DpnII.txt" > "$hic_txt"
+  # Filter BAM file and convert to juicer_short format
+  samtools view --threads 5 -bh -q 30 -f 1 -F 1804 "$bam_file" > "$good_reads_bam"
   if [ $? -ne 0 ]; then
-    echo "Error processing $good_reads_bam with sam2juicer.py."
+    echo "ERROR processing $bam_file with samtools."
+    continue
+  fi
+
+  # Run sam2juicer.py script
+  "$bisulfite_env_python_path" "$software_directory/sam2juicer.py" -s "$good_reads_bam" -f "$fragments_file" > "$hic_txt"
+  if [ $? -ne 0 ]; then
+    echo "WARNING: Problem processing $good_reads_bam with sam2juicer.py. Check the fragments file: $fragments_file"
     continue
   fi
